Initial commit for open-sourced version
This commit contains the initial gud-lldb.el file to be uploaded to GitHub.  It
squashes all the various commits that have accumulated in my private repository
for the last three years while I was both hacking up and using gud-lldb.el on an
almost daily basis.

The code has been used by me primarily on various versions of macOS, starting
with macOS Sierra (10.12), all the way to macOS Catalina (10.15).  The latest
lldb version that I've tested this with is Apple's lldb-1103.0.22.10 from Xcode
version 11.6.

Enjoy and improve.